Select Friends is working with Db errors
Finally, got select friends working. It was a call to getCursor() in
getView() , that is it - one line of code and I swear I hadn't seen
anyone using that in the thoushands of lines of code that I read trying
to get rid of this bug.

There are a lot of warnings about not closing cursor and dbhelper. But,
Sami had figured that out earlier, so hopefully, we will be able to fix
them fast.