SelectFriends working perfectly
Called listView.clearChoices() after overriding @changeCursor in
SelectFriendsAdapter. Another trick was to call
listView.setItemChecked() in the getView() of SelectFriendsAdapter. Both
these were found by @samridhishree

To solve all the database handling problems, we added
GroupBankerApplication which holds a global instance of
FriendsDbAdapter. The change in the manifest is to reflect the new
application class.

Heehah, finally, and hopefully we haven't overlooked anything, we have
the complete package here !